Cleanup and document annotate_known_related*
