+++ oioioi/base/utils/staticfiles.py
@@ -1,0 +1,5 @@
+from django.contrib.staticfiles.storage import ManifestStaticFilesStorage
+
+
+class ManifestLaxStaticFilesStorage(ManifestStaticFilesStorage):
+    manifest_strict = False
